What: Refactor time_code context manager to its own module Why: Utilities should be organized in their own directory How: Create timer.py and import time_code to other files Testing: Works on test PyPI installation
