+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "piano-integration"
-version = "0.0.1"
+version = "0.0.2"
 description = "PIANO: Probabilistic Inference Autoencoder Networks for multi-Omics"
 readme = "README.md"
 requires-python = ">=3.10.18"
