+++ setup.py
@@ -49,7 +49,7 @@ with open('README.md', mode='r') as readme:
 
 setup(
     name='PIANO',
-    version='0.0.1',
+    version='0.0.2',
     packages=find_packages(),
     install_requires=[
         'matplotlib',
